Update game.js
Based on your code, the issue might be occurring due to the event listener for the "Submit Guess" button being set up inside the click event listener of the "Start Game" button. This could be causing multiple event listeners to be attached to the "Submit Guess" button, leading to unintended behavior.

To fix this, we should ensure that the event listener for the "Submit Guess" button is set up independently and only once. Let's restructure your JavaScript code accordingly.